+++ jrnl/jrnl.py
@@ -150,7 +150,12 @@ def cli():
         config = install.install_jrnl(CONFIG_PATH)
     else:
         with open(CONFIG_PATH) as f:
-            config = json.load(f)
+            try:
+                config = json.load(f)
+            except ValueError as e:
+                print("[There seems to be something wrong with your jrnl config at {}: {}]".format(CONFIG_PATH, e.message))
+                print("[Entry was NOT added to your journal]")
+                sys.exit(-1)
         install.update_config(config, config_path=CONFIG_PATH)
 
     original_config = config.copy()